Update app.js
update logic for Bones harvest button: it will now change display state ("Gathering...", "[BONES FULL]", "Process a cadaver") when a Skeleton is summoned.
Will need to extract out the "when to update this button" logic to a separate function at some point so we don't repeat the same "set the UI button again" each time we need to do so.