change Research tooltips to have <hr> to split lore text and bonus description
Add a <hr> tag for all research descriptions to separate the lore blurb and the actual mechanical change.
Instead of displaying the research.description as the content of a div, use vue v-html="research-description" so we can treat the description text as HTML, allowing us to use HTML tags in the text.
We use this functionality to display a horizontal line inside the description text, saving us from needing a separate data property for research.